mutants/json_processor/processor.py

from inspect import signature as _mutmut_signature

def _mutmut_trampoline(orig, mutants, *args, **kwargs):
    import os
    mutant_under_test = os.environ['MUTANT_UNDER_TEST']
    if mutant_under_test == 'fail':
        from mutmut.__main__ import MutmutProgrammaticFailException
        raise MutmutProgrammaticFailException('Failed programmatically')      
    elif mutant_under_test == 'stats':
        from mutmut.__main__ import record_trampoline_hit
        record_trampoline_hit(orig.__module__ + '.' + orig.__name__)
        result = orig(*args, **kwargs)
        return result  # for the yield case
    prefix = orig.__module__ + '.' + orig.__name__ + '__mutmut_'
    if not mutant_under_test.startswith(prefix):
        result = orig(*args, **kwargs)
        return result  # for the yield case
    mutant_name = mutant_under_test.rpartition('.')[-1]
    result = mutants[mutant_name](*args, **kwargs)
    return result


from inspect import signature as _mutmut_signature

def _mutmut_yield_from_trampoline(orig, mutants, *args, **kwargs):
    import os
    mutant_under_test = os.environ['MUTANT_UNDER_TEST']
    if mutant_under_test == 'fail':
        from mutmut.__main__ import MutmutProgrammaticFailException
        raise MutmutProgrammaticFailException('Failed programmatically')      
    elif mutant_under_test == 'stats':
        from mutmut.__main__ import record_trampoline_hit
        record_trampoline_hit(orig.__module__ + '.' + orig.__name__)
        result = yield from orig(*args, **kwargs)
        return result  # for the yield case
    prefix = orig.__module__ + '.' + orig.__name__ + '__mutmut_'
    if not mutant_under_test.startswith(prefix):
        result = yield from orig(*args, **kwargs)
        return result  # for the yield case
    mutant_name = mutant_under_test.rpartition('.')[-1]
    result = yield from mutants[mutant_name](*args, **kwargs)
    return result


import json
import os

class JSONProcessor:
    def xǁJSONProcessorǁ__init____mutmut_orig(self, file_path):
        if not os.path.exists(file_path):
            raise FileNotFoundError(f"File {file_path} not found.")
        self.file_path = file_path
    def xǁJSONProcessorǁ__init____mutmut_1(self, file_path):
        if  os.path.exists(file_path):
            raise FileNotFoundError(f"File {file_path} not found.")
        self.file_path = file_path
    def xǁJSONProcessorǁ__init____mutmut_2(self, file_path):
        if not os.path.exists(None):
            raise FileNotFoundError(f"File {file_path} not found.")
        self.file_path = file_path
    def xǁJSONProcessorǁ__init____mutmut_3(self, file_path):
        if not os.path.exists(file_path):
            raise FileNotFoundError(f"File {file_path} not found.")
        self.file_path = None

    xǁJSONProcessorǁ__init____mutmut_mutants = {
    'xǁJSONProcessorǁ__init____mutmut_1': xǁJSONProcessorǁ__init____mutmut_1, 
        'xǁJSONProcessorǁ__init____mutmut_2': xǁJSONProcessorǁ__init____mutmut_2, 
        'xǁJSONProcessorǁ__init____mutmut_3': xǁJSONProcessorǁ__init____mutmut_3
    }

    def __init__(self, *args, **kwargs):
        result = _mutmut_trampoline(object.__getattribute__(self, "xǁJSONProcessorǁ__init____mutmut_orig"), object.__getattribute__(self, "xǁJSONProcessorǁ__init____mutmut_mutants"), *args, **kwargs)
        return result 

    __init__.__signature__ = _mutmut_signature(xǁJSONProcessorǁ__init____mutmut_orig)
    xǁJSONProcessorǁ__init____mutmut_orig.__name__ = 'xǁJSONProcessorǁ__init__'



    def xǁJSONProcessorǁload_json__mutmut_orig(self):
        with open(self.file_path, 'r') as file:
            return json.load(file)

    def xǁJSONProcessorǁload_json__mutmut_1(self):
        with open(self.file_path, 'XXrXX') as file:
            return json.load(file)

    def xǁJSONProcessorǁload_json__mutmut_2(self):
        with open(self.file_path, 'r') as file:
            return json.load(None)

    xǁJSONProcessorǁload_json__mutmut_mutants = {
    'xǁJSONProcessorǁload_json__mutmut_1': xǁJSONProcessorǁload_json__mutmut_1, 
        'xǁJSONProcessorǁload_json__mutmut_2': xǁJSONProcessorǁload_json__mutmut_2
    }

    def load_json(self, *args, **kwargs):
        result = _mutmut_trampoline(object.__getattribute__(self, "xǁJSONProcessorǁload_json__mutmut_orig"), object.__getattribute__(self, "xǁJSONProcessorǁload_json__mutmut_mutants"), *args, **kwargs)
        return result 

    load_json.__signature__ = _mutmut_signature(xǁJSONProcessorǁload_json__mutmut_orig)
    xǁJSONProcessorǁload_json__mutmut_orig.__name__ = 'xǁJSONProcessorǁload_json'



    def xǁJSONProcessorǁvalidate_json__mutmut_orig(self, schema):
        data = self.load_json()
        return all(key in data for key in schema)

    def xǁJSONProcessorǁvalidate_json__mutmut_1(self, schema):
        data = None
        return all(key in data for key in schema)

    def xǁJSONProcessorǁvalidate_json__mutmut_2(self, schema):
        data = self.load_json()
        return all(key not in data for key in schema)

    xǁJSONProcessorǁvalidate_json__mutmut_mutants = {
    'xǁJSONProcessorǁvalidate_json__mutmut_1': xǁJSONProcessorǁvalidate_json__mutmut_1, 
        'xǁJSONProcessorǁvalidate_json__mutmut_2': xǁJSONProcessorǁvalidate_json__mutmut_2
    }

    def validate_json(self, *args, **kwargs):
        result = _mutmut_trampoline(object.__getattribute__(self, "xǁJSONProcessorǁvalidate_json__mutmut_orig"), object.__getattribute__(self, "xǁJSONProcessorǁvalidate_json__mutmut_mutants"), *args, **kwargs)
        return result 

    validate_json.__signature__ = _mutmut_signature(xǁJSONProcessorǁvalidate_json__mutmut_orig)
    xǁJSONProcessorǁvalidate_json__mutmut_orig.__name__ = 'xǁJSONProcessorǁvalidate_json'



    def xǁJSONProcessorǁmodify_json__mutmut_orig(self, key, value):
        data = self.load_json()
        data[key] = value
        with open(self.file_path, 'w') as file:
            json.dump(data, file, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_1(self, key, value):
        data = None
        data[key] = value
        with open(self.file_path, 'w') as file:
            json.dump(data, file, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_2(self, key, value):
        data = self.load_json()
        data[None] = value
        with open(self.file_path, 'w') as file:
            json.dump(data, file, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_3(self, key, value):
        data = self.load_json()
        data[key] = None
        with open(self.file_path, 'w') as file:
            json.dump(data, file, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_4(self, key, value):
        data = self.load_json()
        data[key] = value
        with open(self.file_path, 'XXwXX') as file:
            json.dump(data, file, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_5(self, key, value):
        data = self.load_json()
        data[key] = value
        with open(self.file_path, 'w') as file:
            json.dump(None, file, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_6(self, key, value):
        data = self.load_json()
        data[key] = value
        with open(self.file_path, 'w') as file:
            json.dump(data, None, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_7(self, key, value):
        data = self.load_json()
        data[key] = value
        with open(self.file_path, 'w') as file:
            json.dump(data, file, indent=5)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_8(self, key, value):
        data = self.load_json()
        data[key] = value
        with open(self.file_path, 'w') as file:
            json.dump( file, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_9(self, key, value):
        data = self.load_json()
        data[key] = value
        with open(self.file_path, 'w') as file:
            json.dump(data, indent=4)
        return data

    def xǁJSONProcessorǁmodify_json__mutmut_10(self, key, value):
        data = self.load_json()
        data[key] = value
        with open(self.file_path, 'w') as file:
            json.dump(data, file,)
        return data

    xǁJSONProcessorǁmodify_json__mutmut_mutants = {
    'xǁJSONProcessorǁmodify_json__mutmut_1': xǁJSONProcessorǁmodify_json__mutmut_1, 
        'xǁJSONProcessorǁmodify_json__mutmut_2': xǁJSONProcessorǁmodify_json__mutmut_2, 
        'xǁJSONProcessorǁmodify_json__mutmut_3': xǁJSONProcessorǁmodify_json__mutmut_3, 
        'xǁJSONProcessorǁmodify_json__mutmut_4': xǁJSONProcessorǁmodify_json__mutmut_4, 
        'xǁJSONProcessorǁmodify_json__mutmut_5': xǁJSONProcessorǁmodify_json__mutmut_5, 
        'xǁJSONProcessorǁmodify_json__mutmut_6': xǁJSONProcessorǁmodify_json__mutmut_6, 
        'xǁJSONProcessorǁmodify_json__mutmut_7': xǁJSONProcessorǁmodify_json__mutmut_7, 
        'xǁJSONProcessorǁmodify_json__mutmut_8': xǁJSONProcessorǁmodify_json__mutmut_8, 
        'xǁJSONProcessorǁmodify_json__mutmut_9': xǁJSONProcessorǁmodify_json__mutmut_9, 
        'xǁJSONProcessorǁmodify_json__mutmut_10': xǁJSONProcessorǁmodify_json__mutmut_10
    }

    def modify_json(self, *args, **kwargs):
        result = _mutmut_trampoline(object.__getattribute__(self, "xǁJSONProcessorǁmodify_json__mutmut_orig"), object.__getattribute__(self, "xǁJSONProcessorǁmodify_json__mutmut_mutants"), *args, **kwargs)
        return result 

    modify_json.__signature__ = _mutmut_signature(xǁJSONProcessorǁmodify_json__mutmut_orig)
    xǁJSONProcessorǁmodify_json__mutmut_orig.__name__ = 'xǁJSONProcessorǁmodify_json'


